refactor(server): isolate statistics pandas implem
